improvement: Update service controller Change file name
